+++ Cache.hpp
@@ -1,27 +1,27 @@
 class Cache {
     // deals with the LRU
-    unsigned int associativity;
-    unsigned int blockSize;
-    unsigned int cacheSize;
+    size_t associativity;
+    size_t blockSize;
+    size_t cacheSize;
     std::vector<CacheSet> cache;
     // the cache is being split into tag, index and offset
     // offset --> blockSize
     // index --> location of the block
     // tag --> is the associated memory stored in the cache
     public:
-        Cache(unsigned int cacheSize, unsigned int associativity, unsigned int blockSize){};
+        Cache(size_t cacheSize, size_t associativity, size_t blockSize){};
 
-        unsigned int getIndex(unsigned int address){};
+        size_t getIndex(size_t address){};
 
-        unsigned int getTag(unsigned int address){};
+        size_t getTag(size_t address){};
 
-        bool checkCacheLine(unsigned int address){};
+        bool checkCacheLine(size_t address){};
 
-        void invalidateCacheLine(unsigned int address){};
+        void invalidateCacheLine(size_t address){};
 
-        void addCacheLine(unsigned int address, unsigned int state){};
+        void addCacheLine(size_t address, size_t state){};
 
-        bool readCacheLine(unsigned int address){};
+        bool readCacheLine(size_t address){};
 
-        bool updateCacheLine(unsigned int address,unsigned int state){};
+        bool updateCacheLine(size_t address,size_t state){};
 };
